Add ingredient storage system and fix calories sum
Ingredient count for a DishPart is stored in the DishPartBase object
in a Dictionary that holds the IDishPart derived object and a count
of needed ingredients for a dish.

The function that counts calories of all ingredients and other
DishParts is now working with different types of DishPartBase
derived objects present in the DishPart.