Add creating of entities in a concole interface
Implemented an options handler that interacts with the user using
the letters of an alphabet. These letter are asigned to strings
describing what to do that are mapped to functions.

Implemented adding functions for meals and ingredients using a
generic method that calls for each property a mapped function
that fills the property.